Add support for timeouts in HTTPTransport
Use HTTPTransport.setTimeout(float) for setting timeouts, just like for
PBCTransport. Calls to Agent.request that exceed the timeout will raise
a TimeoutError exception. This is different from PBCTransport which at the
moment raises a generic RiakPBCException.

It would probably be better to wrap RiakPBCException in TimeoutError
when appropriate, so that both transports share the same interface.