Fix PBC problem with transports hanging on connection refused
If a connection attempt is refused, Twisted calls the
clientConnectionFailed method of the client factory to determine what to
do. The default behaviour is to do nothing, thereby hanging the
connection. This patch changes the behaviour of RiakPBCClientFactory so
that the error is passed back to whoever initiated the connection in the
first place.